+++ src/Field.php
@@ -324,6 +324,7 @@ class Field extends BaseField implements EagerLoadingFieldInterface, GqlInlineFr
                 $newBlockType->name = $blockType['name'];
                 $newBlockType->handle = $blockType['handle'];
                 $newBlockType->enabled = $blockType['enabled'];
+                $newBlockType->ignorePermissions = $blockType['ignorePermissions'];
                 $newBlockType->description = $blockType['description'];
                 $newBlockType->minBlocks = (int)$blockType['minBlocks'];
                 $newBlockType->maxBlocks = (int)$blockType['maxBlocks'];