+++ src/Plugin.php
@@ -59,7 +59,7 @@ class Plugin extends BasePlugin
     /**
      * @inheritdoc
      */
-    public string $schemaVersion = '3.3.0';
+    public string $schemaVersion = '3.5.2';
 
     /**
      * @inheritdoc