+++ src/translations/en/neo.php
@@ -41,6 +41,8 @@ return [
     'The maximum number of child blocks this block type is allowed to have.' => 'The maximum number of child blocks this block type is allowed to have.',
     'Top Level' => 'Top Level',
     'Will this block type be allowed at the top level?' => 'Will this block type be allowed at the top level?',
+    'Ignore Permissions' => 'Ignore Permissions',
+    'Whether user permissions for this block type should be ignored.' => 'Whether user permissions for this block type should be ignored.',
     'Group Child Block Types' => 'Group Child Block Types',
     'Will child block types be shown in their groups?' => 'Will child block types be shown in their groups?',
     '{type} Condition' => '{type} Condition',